Effekt-specific shell options
This commit introduces options for an Effekt-specific shell. In the extension settings you will find options for a path to the desired shell binary and for the path to the "/bin"-folder of Effekt (current directory of the "effekt"-bash script).
he extension now holds the command "Effekt: Effekt Terminal". If the chosen shell supports bash shell scripts, the extension adds the path of the bash script to the PATH-variable. Thus, issueing the command "effekt <args>" in the Effekt Terminal will call the right shell script. Furthermore, the current working directory of the newly opened shell is set to the folder of the currently opened file (misbehaves if no folder/file is opened in VSCode!).
Current status is probably NOT platform independent.